requirements.txt
gensim
jieba
logger
numpy
requests
pyemd
pyahocorasick
numpy
requests
tensorflow==1.9.0